Prepare video visualizer for upcoming stream visualizer * rename `VideoCache` to `VideoAssetCache` * move out utilities from video frame reference visualizer to be used in video stream * handle pickable rect bounding boxes more uniformly
