+++ crates/viewer/re_view_spatial/src/visualizers/mod.rs
@@ -19,7 +19,7 @@ mod points3d;
 mod segmentation_images;
 mod transform3d_arrows;
 mod utilities;
-mod videos;
+mod video;
 
 pub use cameras::CamerasVisualizer;
 pub use depth_images::DepthImageVisualizer;
@@ -85,7 +85,7 @@ pub fn register_2d_spatial_visualizers(
     system_registry.register_visualizer::<segmentation_images::SegmentationImageVisualizer>()?;
     system_registry.register_visualizer::<transform3d_arrows::AxisLengthDetector>()?;
     system_registry.register_visualizer::<transform3d_arrows::Transform3DArrowsVisualizer>()?;
-    system_registry.register_visualizer::<videos::VideoFrameReferenceVisualizer>()?;
+    system_registry.register_visualizer::<video::VideoFrameReferenceVisualizer>()?;
     Ok(())
 }
 
@@ -97,9 +97,10 @@ pub fn register_3d_spatial_visualizers(
     system_registry.register_visualizer::<assets3d::Asset3DVisualizer>()?;
     system_registry.register_visualizer::<boxes2d::Boxes2DVisualizer>()?;
     system_registry.register_visualizer::<boxes3d::Boxes3DVisualizer>()?;
-    system_registry.register_visualizer::<capsules3d::Capsules3DVisualizer>()?;
     system_registry.register_visualizer::<cameras::CamerasVisualizer>()?;
+    system_registry.register_visualizer::<capsules3d::Capsules3DVisualizer>()?;
     system_registry.register_visualizer::<depth_images::DepthImageVisualizer>()?;
+    system_registry.register_visualizer::<ellipsoids::Ellipsoids3DVisualizer>()?;
     system_registry.register_visualizer::<encoded_image::EncodedImageVisualizer>()?;
     system_registry.register_visualizer::<images::ImageVisualizer>()?;
     system_registry.register_visualizer::<lines2d::Lines2DVisualizer>()?;
@@ -108,10 +109,9 @@ pub fn register_3d_spatial_visualizers(
     system_registry.register_visualizer::<points2d::Points2DVisualizer>()?;
     system_registry.register_visualizer::<points3d::Points3DVisualizer>()?;
     system_registry.register_visualizer::<segmentation_images::SegmentationImageVisualizer>()?;
-    system_registry.register_visualizer::<ellipsoids::Ellipsoids3DVisualizer>()?;
     system_registry.register_visualizer::<transform3d_arrows::AxisLengthDetector>()?;
     system_registry.register_visualizer::<transform3d_arrows::Transform3DArrowsVisualizer>()?;
-    system_registry.register_visualizer::<videos::VideoFrameReferenceVisualizer>()?;
+    system_registry.register_visualizer::<video::VideoFrameReferenceVisualizer>()?;
     Ok(())
 }
 
@@ -153,7 +153,7 @@ pub fn visualizers_processing_draw_order()
             archetypes::SegmentationImage::descriptor_draw_order(),
         ),
         (
-            videos::VideoFrameReferenceVisualizer::identifier(),
+            video::VideoFrameReferenceVisualizer::identifier(),
             archetypes::VideoFrameReference::descriptor_draw_order(),
         ),
     ]
