+++ crates/viewer/re_view_spatial/src/visualizers/segmentation_images.rs
@@ -127,16 +127,18 @@ impl VisualizerSystem for SegmentationImageVisualizer {
                         colormap,
                         multiplicative_tint,
                         SegmentationImage::name(),
-                        &mut self.data,
                     ) {
-                        self.data.pickable_rects.push(PickableTexturedRect {
-                            ent_path: entity_path.clone(),
-                            textured_rect,
-                            source_data: PickableRectSourceData::Image {
-                                image,
-                                depth_meter: None,
+                        self.data.add_pickable_rect(
+                            PickableTexturedRect {
+                                ent_path: entity_path.clone(),
+                                textured_rect,
+                                source_data: PickableRectSourceData::Image {
+                                    image,
+                                    depth_meter: None,
+                                },
                             },
-                        });
+                            spatial_ctx.view_class_identifier,
+                        );
                     }
                 }
