+++ crates/viewer/re_view_spatial/src/visualizers/utilities/spatial_view_visualizer.rs
@@ -1,7 +1,11 @@
 use re_log_types::EntityPathHash;
+use re_types::ViewClassIdentifier;
+use re_viewer_context::ViewClass as _;
 
 use super::UiLabel;
-use crate::{PickableTexturedRect, view_kind::SpatialViewKind, visualizers::LoadingSpinner};
+use crate::{
+    PickableTexturedRect, SpatialView2D, view_kind::SpatialViewKind, visualizers::LoadingSpinner,
+};
 
 /// Common data struct for all spatial scene elements.
 ///
@@ -34,6 +38,15 @@ impl SpatialViewVisualizerData {
         }
     }
 
+    pub fn add_pickable_rect(
+        &mut self,
+        pickable_rect: PickableTexturedRect,
+        class_identifier: ViewClassIdentifier,
+    ) {
+        self.add_pickable_rect_to_bounding_box(&pickable_rect, class_identifier);
+        self.pickable_rects.push(pickable_rect);
+    }
+
     pub fn add_bounding_box(
         &mut self,
         entity: EntityPathHash,
@@ -42,6 +55,23 @@ impl SpatialViewVisualizerData {
     ) {
         self.bounding_boxes
             .push((entity, bbox.transform_affine3(&world_from_obj)));
+    }
+
+    pub fn add_pickable_rect_to_bounding_box(
+        &mut self,
+        pickable_rect: &PickableTexturedRect,
+        class_identifier: ViewClassIdentifier,
+    ) {
+        // Only update the bounding box if this is a 2D view.
+        // This is avoids a cyclic relationship where the image plane grows
+        // the bounds which in turn influence the size of the image plane.
+        // See: https://github.com/rerun-io/rerun/issues/3728
+        if class_identifier == SpatialView2D::identifier() {
+            self.bounding_boxes.push((
+                pickable_rect.ent_path.hash(),
+                bounding_box_for_textured_rect(&pickable_rect.textured_rect),
+            ));
+        }
     }
 
     pub fn add_bounding_box_from_points(
@@ -61,4 +91,22 @@ impl SpatialViewVisualizerData {
     pub fn as_any(&self) -> &dyn std::any::Any {
         self
     }
+}
+
+fn bounding_box_for_textured_rect(
+    textured_rect: &re_renderer::renderer::TexturedRect,
+) -> macaw::BoundingBox {
+    let left_top = textured_rect.top_left_corner_position;
+    let extent_u = textured_rect.extent_u;
+    let extent_v = textured_rect.extent_v;
+
+    macaw::BoundingBox::from_points(
+        [
+            left_top,
+            left_top + extent_u,
+            left_top + extent_v,
+            left_top + extent_v + extent_u,
+        ]
+        .into_iter(),
+    )
 }
