+++ crates/viewer/re_view_spatial/src/visualizers/utilities/textured_rect.rs
@@ -3,13 +3,9 @@ use glam::Vec3;
 use re_log_types::EntityPath;
 use re_renderer::renderer;
 use re_types::ArchetypeName;
-use re_viewer_context::{
-    ColormapWithRange, ImageInfo, ImageStatsCache, ViewClass as _, ViewerContext, gpu_bridge,
-};
+use re_viewer_context::{ColormapWithRange, ImageInfo, ImageStatsCache, ViewerContext, gpu_bridge};
 
-use crate::{SpatialView2D, contexts::SpatialSceneEntityContext};
-
-use super::SpatialViewVisualizerData;
+use crate::contexts::SpatialSceneEntityContext;
 
 #[allow(clippy::too_many_arguments)]
 pub fn textured_rect_from_image(
@@ -20,7 +16,6 @@ pub fn textured_rect_from_image(
     colormap: Option<&ColormapWithRange>,
     multiplicative_tint: egui::Rgba,
     archetype_name: ArchetypeName,
-    visualizer_data: &mut SpatialViewVisualizerData,
 ) -> Option<renderer::TexturedRect> {
     let debug_name = ent_path.to_string();
     let tensor_stats = ctx
@@ -74,18 +69,6 @@ pub fn textured_rect_from_image(
                 },
             };
 
-            // Only update the bounding box if this is a 2D view.
-            // This is avoids a cyclic relationship where the image plane grows
-            // the bounds which in turn influence the size of the image plane.
-            // See: https://github.com/rerun-io/rerun/issues/3728
-            if ent_context.view_class_identifier == SpatialView2D::identifier() {
-                visualizer_data.add_bounding_box(
-                    ent_path.hash(),
-                    bounding_box_for_textured_rect(&textured_rect),
-                    world_from_entity,
-                );
-            }
-
             Some(textured_rect)
         }
 
@@ -94,20 +77,4 @@ pub fn textured_rect_from_image(
             None
         }
     }
-}
-
-fn bounding_box_for_textured_rect(textured_rect: &renderer::TexturedRect) -> macaw::BoundingBox {
-    let left_top = textured_rect.top_left_corner_position;
-    let extent_u = textured_rect.extent_u;
-    let extent_v = textured_rect.extent_v;
-
-    macaw::BoundingBox::from_points(
-        [
-            left_top,
-            left_top + extent_u,
-            left_top + extent_v,
-            left_top + extent_v + extent_u,
-        ]
-        .into_iter(),
-    )
 }
